# improve method `cycle_basis` for graphs with multiple edges