@kwankyu 2023-10-24
Looks good to me though I don't know the algorithm.

Can you provide some benchmarks showing that the new algorithm is more efficient?