@kwankyu 2023-10-25
If you want to attract attention from graph theory experts for more time, you may discard my review.