@dcoudert 2023-10-31
This is fixed in #36609